finito save e shop system, aggiunto PlayerData

+++ Assets/_Project/Scripts/SaveManager/DataContainer.cs
@@ -2,7 +2,7 @@ public class DataContainer
 {
     public int Coins;
     public int HealthUp_count;
-    public float SpeedBoost_count;
+    public int SpeedBoost_count;
     public int Shield_count;
     public bool EnableDoubleJump_purchased;
     public bool DoubleCoin_purchased;
